Split _thrreaded internal module to 3 modules: 1 helper = 1 module
Make 2.0.1
