+++ threaded/_threaded.py
@@ -12,188 +12,21 @@
 #    License for the specific language governing permissions and limitations
 #    under the License.
 
-"""Python 3 threaded implementation.
+"""Threaded implementation.
 
 Asyncio is supported
 """
 
-import asyncio
-import concurrent.futures
 import functools
 import threading
 import typing
 
-from . import _base_threaded
 from . import _class_decorator
 
 __all__ = (
-    'ThreadPooled',
     'Threaded',
-    'AsyncIOTask',
-    'threadpooled',
     'threaded',
-    'asynciotask',
 )
-
-
-def _await_if_required(target: typing.Callable) -> typing.Callable[..., typing.Any]:
-    """Await result if coroutine was returned."""
-    @functools.wraps(target)
-    def wrapper(
-        *args: typing.Tuple,
-        **kwargs: typing.Dict
-    ) -> typing.Any:
-        """Decorator/wrapper."""
-        result = target(*args, **kwargs)
-        if asyncio.iscoroutine(result):
-            loop = asyncio.new_event_loop()
-            result = loop.run_until_complete(result)
-            loop.close()
-        return result
-
-    return wrapper
-
-
-class ThreadPooled(_base_threaded.BasePooled):
-    """Post function to ThreadPoolExecutor."""
-
-    __slots__ = (
-        '__loop_getter',
-        '__loop_getter_need_context'
-    )
-
-    def __init__(
-        self,
-        func: typing.Optional[typing.Callable] = None,
-        *,
-        loop_getter: typing.Optional[
-            typing.Union[
-                typing.Callable[..., asyncio.AbstractEventLoop],
-                asyncio.AbstractEventLoop
-            ]
-        ] = None,
-        loop_getter_need_context: bool = False
-    ) -> None:
-        """Wrap function in future and return.
-
-        :param func: function to wrap
-        :type func: typing.Optional[typing.Callable]
-        :param loop_getter: Method to get event loop, if wrap in asyncio task
-        :type loop_getter: typing.Union[
-                               None,
-                               typing.Callable[..., asyncio.AbstractEventLoop],
-                               asyncio.AbstractEventLoop
-                           ]
-        :param loop_getter_need_context: Loop getter requires function context
-        :type loop_getter_need_context: bool
-        """
-        super(ThreadPooled, self).__init__(func=func)
-        self.__loop_getter = loop_getter
-        self.__loop_getter_need_context = loop_getter_need_context
-
-    @property
-    def loop_getter(
-        self
-    ) -> typing.Optional[
-        typing.Union[
-            typing.Callable[..., asyncio.AbstractEventLoop],
-            asyncio.AbstractEventLoop
-        ]
-    ]:
-        """Loop getter.
-
-        :rtype: typing.Union[
-                    None,
-                    typing.Callable[..., asyncio.AbstractEventLoop],
-                    asyncio.AbstractEventLoop
-                ]
-        """
-        return self.__loop_getter
-
-    @property
-    def loop_getter_need_context(self) -> bool:
-        """Loop getter need execution context.
-
-        :rtype: bool
-        """
-        return self.__loop_getter_need_context
-
-    def _get_loop(
-        self,
-        *args: typing.Tuple,
-        **kwargs: typing.Dict
-    ) -> typing.Optional[asyncio.AbstractEventLoop]:
-        """Get event loop in decorator class."""
-        if callable(self.loop_getter):
-            if self.loop_getter_need_context:
-                return self.loop_getter(*args, **kwargs)  # pylint: disable=not-callable
-            return self.loop_getter()  # pylint: disable=not-callable
-        return self.loop_getter
-
-    def _get_function_wrapper(
-        self,
-        func: typing.Callable
-    ) -> typing.Callable[..., typing.Union[typing.Awaitable, concurrent.futures.Future]]:
-        """Here should be constructed and returned real decorator.
-
-        :param func: Wrapped function
-        :type func: typing.Callable
-        :return: wrapped coroutine or function
-        :rtype: typing.Callable[..., typing.Union[typing.Awaitable, concurrent.futures.Future]]
-        """
-        prepared = _await_if_required(func)
-
-        # pylint: disable=missing-docstring
-        # noinspection PyMissingOrEmptyDocstring
-        @functools.wraps(prepared)
-        def wrapper(
-            *args: typing.Tuple,
-            **kwargs: typing.Dict
-        ) -> typing.Union[
-            typing.Awaitable, concurrent.futures.Future,
-            typing.Callable[..., typing.Union[typing.Awaitable, concurrent.futures.Future]]
-        ]:
-            loop = self._get_loop(*args, **kwargs)  # type: ignore
-
-            if loop is None:
-                return self.executor.submit(prepared, *args, **kwargs)
-
-            return loop.run_in_executor(
-                self.executor,
-                functools.partial(
-                    prepared,
-                    *args, **kwargs
-                )
-            )
-
-        # pylint: enable=missing-docstring
-        return wrapper
-
-    def __call__(  # pylint: disable=useless-super-delegation
-        self,
-        *args: typing.Union[typing.Tuple, typing.Callable],
-        **kwargs: typing.Dict
-    ) -> typing.Union[
-        concurrent.futures.Future, typing.Awaitable,
-        typing.Callable[..., typing.Union[typing.Awaitable, concurrent.futures.Future]]
-    ]:
-        """Callable instance."""
-        return super(ThreadPooled, self).__call__(*args, **kwargs)  # type: ignore
-
-    def __repr__(self) -> str:
-        """For debug purposes."""
-        return (
-            "<{cls}("
-            "{func!r}, "
-            "loop_getter={self.loop_getter!r}, "
-            "loop_getter_need_context={self.loop_getter_need_context!r}, "
-            ") at 0x{id:X}>".format(
-                cls=self.__class__.__name__,
-                func=self._func,
-                self=self,
-                id=id(self)
-            )
-        )  # pragma: no cover
 
 
 class Threaded(_class_decorator.BaseDecorator):
@@ -281,7 +114,7 @@ class Threaded(_class_decorator.BaseDecorator):
         :return: wrapped function
         :rtype: typing.Callable[..., threading.Thread]
         """
-        prepared = _await_if_required(func)
+        prepared = self._await_if_required(func)
         name = self.name
         if name is None:
             name = 'Threaded: ' + getattr(
@@ -320,207 +153,7 @@ class Threaded(_class_decorator.BaseDecorator):
         return super(Threaded, self).__call__(*args, **kwargs)  # type: ignore
 
 
-class AsyncIOTask(_class_decorator.BaseDecorator):
-    """Wrap to asyncio.Task."""
-
-    __slots__ = (
-        '__loop_getter',
-        '__loop_getter_need_context',
-    )
-
-    def __init__(
-        self,
-        func: typing.Optional[typing.Callable] = None,
-        *,
-        loop_getter: typing.Union[
-            typing.Callable[..., asyncio.AbstractEventLoop],
-            asyncio.AbstractEventLoop
-        ] = asyncio.get_event_loop,
-        loop_getter_need_context: bool = False
-    ) -> None:
-        """Wrap function in future and return.
-
-        :param func: Function to wrap
-        :type func: typing.Optional[typing.Callable]
-        :param loop_getter: Method to get event loop, if wrap in asyncio task
-        :type loop_getter: typing.Union[
-                               typing.Callable[..., asyncio.AbstractEventLoop],
-                               asyncio.AbstractEventLoop
-                           ]
-        :param loop_getter_need_context: Loop getter requires function context
-        :type loop_getter_need_context: bool
-        """
-        super(AsyncIOTask, self).__init__(func=func)
-        self.__loop_getter = loop_getter
-        self.__loop_getter_need_context = loop_getter_need_context
-
-    @property
-    def loop_getter(
-            self
-    ) -> typing.Union[
-        typing.Callable[..., asyncio.AbstractEventLoop],
-        asyncio.AbstractEventLoop
-    ]:
-        """Loop getter.
-
-        :rtype: typing.Union[
-                    typing.Callable[..., asyncio.AbstractEventLoop],
-                    asyncio.AbstractEventLoop
-                ]
-        """
-        return self.__loop_getter
-
-    @property
-    def loop_getter_need_context(self) -> bool:
-        """Loop getter need execution context.
-
-        :rtype: bool
-        """
-        return self.__loop_getter_need_context
-
-    def get_loop(
-        self,
-        *args,  # type: typing.Tuple
-        **kwargs  # type: typing.Dict
-    ) -> asyncio.AbstractEventLoop:
-        """Get event loop in decorator class."""
-        if callable(self.loop_getter):
-            if self.loop_getter_need_context:
-                return self.loop_getter(*args, **kwargs)  # pylint: disable=not-callable
-            return self.loop_getter()  # pylint: disable=not-callable
-        return self.loop_getter
-
-    def _get_function_wrapper(
-        self,
-        func: typing.Callable
-    ) -> typing.Callable[..., asyncio.Task]:
-        """Here should be constructed and returned real decorator.
-
-        :param func: Wrapped function
-        :type func: typing.Callable
-        :rtype: typing.Callable[..., asyncio.Task]
-        """
-        # pylint: disable=missing-docstring
-        # noinspection PyMissingOrEmptyDocstring
-        @functools.wraps(func)
-        def wrapper(
-            *args,  # type: typing.Tuple
-            **kwargs  # type: typing.Dict
-        ) -> asyncio.Task:
-            loop = self.get_loop(*args, **kwargs)  # type: ignore
-            return loop.create_task(func(*args, **kwargs))
-
-        # pylint: enable=missing-docstring
-        return wrapper
-
-    def __call__(  # pylint: disable=useless-super-delegation
-        self,
-        *args: typing.Union[typing.Tuple, typing.Callable],
-        **kwargs: typing.Dict
-    ) -> typing.Union[asyncio.Task, typing.Callable[..., asyncio.Task]]:
-        """Callable instance."""
-        return super(AsyncIOTask, self).__call__(*args, **kwargs)  # type: ignore
-
-    def __repr__(self) -> str:
-        """For debug purposes."""
-        return (
-            "<{cls}("
-            "{func!r}, "
-            "loop_getter={self.loop_getter!r}, "
-            "loop_getter_need_context={self.loop_getter_need_context!r}, "
-            ") at 0x{id:X}>".format(
-                cls=self.__class__.__name__,
-                func=self._func,
-                self=self,
-                id=id(self)
-            )
-        )  # pragma: no cover
-
-
 # pylint: disable=function-redefined, unused-argument
-@typing.overload
-def threadpooled(
-    func: typing.Callable,
-    *,
-    loop_getter: None = None,
-    loop_getter_need_context: bool = False
-) -> typing.Callable[..., concurrent.futures.Future]:
-    """Overload: function callable, no loop getter."""
-    pass  # pragma: no cover
-
-
-@typing.overload  # noqa: F811
-def threadpooled(
-    func: typing.Callable,
-    *,
-    loop_getter: typing.Union[
-        typing.Callable[..., asyncio.AbstractEventLoop],
-        asyncio.AbstractEventLoop
-    ],
-    loop_getter_need_context: bool = False
-) -> typing.Callable[..., asyncio.Task]:
-    """Overload: function callable, loop getter available."""
-    pass  # pragma: no cover
-
-
-@typing.overload  # noqa: F811
-def threadpooled(
-    func: None = None,
-    *,
-    loop_getter: typing.Union[
-        None,
-        typing.Callable[..., asyncio.AbstractEventLoop],
-        asyncio.AbstractEventLoop
-    ] = None,
-    loop_getter_need_context: bool = False
-) -> ThreadPooled:
-    """Overload: No function."""
-    pass  # pragma: no cover
-
-
-# pylint: enable=unused-argument
-# pylint: disable=unexpected-keyword-arg, no-value-for-parameter
-def threadpooled(  # noqa: F811
-    func: typing.Optional[typing.Callable] = None,
-    *,
-    loop_getter: typing.Union[
-        None,
-        typing.Callable[..., asyncio.AbstractEventLoop],
-        asyncio.AbstractEventLoop
-    ] = None,
-    loop_getter_need_context: bool = False
-) -> typing.Union[
-    ThreadPooled,
-    typing.Callable[..., typing.Union[concurrent.futures.Future, typing.Awaitable]]
-]:
-    """Post function to ThreadPoolExecutor.
-
-    :param func: function to wrap
-    :type func: typing.Optional[typing.Callable]
-    :param loop_getter: Method to get event loop, if wrap in asyncio task
-    :type loop_getter: typing.Union[
-                           None,
-                           typing.Callable[..., asyncio.AbstractEventLoop],
-                           asyncio.AbstractEventLoop
-                       ]
-    :param loop_getter_need_context: Loop getter requires function context
-    :type loop_getter_need_context: bool
-    :rtype: typing.Union[ThreadPooled, typing.Callable[..., typing.Union[concurrent.futures.Future, typing.Awaitable]]]
-    """
-    if func is None:
-        return ThreadPooled(
-            func=func,
-            loop_getter=loop_getter,
-            loop_getter_need_context=loop_getter_need_context
-        )
-    return ThreadPooled(  # type: ignore
-        func=None,
-        loop_getter=loop_getter,
-        loop_getter_need_context=loop_getter_need_context
-    )(func)
-
-
-# pylint: disable=unused-argument
 @typing.overload
 def threaded(
     name: typing.Callable,
@@ -566,69 +199,4 @@ def threaded(  # noqa: F811
         )
         return Threaded(name=name, daemon=daemon, started=started)(func)  # type: ignore
     return Threaded(name=name, daemon=daemon, started=started)
-
-
-# pylint: disable=unused-argument
-@typing.overload
-def asynciotask(
-    func: None = None,
-    *,
-    loop_getter: typing.Union[
-        typing.Callable[..., asyncio.AbstractEventLoop],
-        asyncio.AbstractEventLoop
-    ] = asyncio.get_event_loop,
-    loop_getter_need_context: bool = False
-) -> AsyncIOTask:
-    """Overload: no function."""
-    pass  # pragma: no cover
-
-
-@typing.overload  # noqa: F811
-def asynciotask(
-    func: typing.Callable,
-    *,
-    loop_getter: typing.Union[
-        typing.Callable[..., asyncio.AbstractEventLoop],
-        asyncio.AbstractEventLoop
-    ] = asyncio.get_event_loop,
-    loop_getter_need_context: bool = False
-) -> typing.Callable[..., asyncio.Task]:
-    """Overload: provided function."""
-    pass  # pragma: no cover
-
-
-# pylint: enable=unused-argument
-def asynciotask(  # noqa: F811
-    func: typing.Optional[typing.Callable] = None,
-    *,
-    loop_getter: typing.Union[
-        typing.Callable[..., asyncio.AbstractEventLoop],
-        asyncio.AbstractEventLoop
-    ] = asyncio.get_event_loop,
-    loop_getter_need_context: bool = False
-) -> typing.Union[AsyncIOTask, typing.Callable[..., asyncio.Task]]:
-    """Wrap function in future and return.
-
-    :param func: Function to wrap
-    :type func: typing.Optional[typing.Callable]
-    :param loop_getter: Method to get event loop, if wrap in asyncio task
-    :type loop_getter: typing.Union[
-                           typing.Callable[..., asyncio.AbstractEventLoop],
-                           asyncio.AbstractEventLoop
-                       ]
-    :param loop_getter_need_context: Loop getter requires function context
-    :type loop_getter_need_context: bool
-    :rtype: typing.Union[AsyncIOTask, typing.Callable[..., asyncio.Task]]
-    """
-    if func is None:
-        return AsyncIOTask(
-            func=func,
-            loop_getter=loop_getter,
-            loop_getter_need_context=loop_getter_need_context
-        )
-    return AsyncIOTask(  # type: ignore
-        func=None,
-        loop_getter=loop_getter,
-        loop_getter_need_context=loop_getter_need_context
-    )(func)
 # pylint: enable=unexpected-keyword-arg, no-value-for-parameter, function-redefined
